baseline-mem0/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2.0
pydantic-settings>=2.0
mem0ai>=0.1.0
httpx>=0.27
